extract common parsing logics
